Create code.m
Continuous Stiffness Measurement (CSM) is a technique used during nanoindentation experiments, where the stiffness is continuously measured as a function of indentation depth. This allows obtaining the indentation modulus and hardness as a function of depth. These depth-dependent properties can then be used to derive the stress-strain curve, considering the material's response at different indentation depths.